+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Only listing significant user-visible, not internal code cleanups and minor bug 
 * [QATM-3284] Fix loop replacement in datatables
 * Add base service clients and models
 * [QATM-3353] Fix createResourceIfNotExist method
+* [QATM-3375] New step to obtain registered version of a service in gosec management
 
 ## 0.17.0 (May 25, 2020)